requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.6.1
pydantic==2.9.2
python-dotenv==1.0.1